requirements.txt
streamlit>=1.28.0
google-generativeai>=0.5.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
pillow>=10.0.0
opencv-python>=4.8.0
moviepy>=1.0.3 